Successfully extended the Object class and the generateJSON function to allow for custom attributes.
Improvements to loading JSON - linking objects to their lines and connectors.
Changed to DatalessJson using svg paths.
Still have problems with the applying the extensions to the path-group (SVG) classes

iconate stuff probably will not be used - for animating icons on the page (i.e. fade in/out), don't think it links to canvas. Better to use this method built into Fabric:
https://github.com/kangax/fabric.js/issues/77